Makefile made more efficient
The makefile was using directories as dependencies which led to inconsistent and
unneeded recompilation of source code. Now the makefile will only recompile
files that have changed.

One major issue still needing resolved is that the makefile does not detect
dependencies of the source code, at this time I do not see a clear way to fix
this.