EJS pages now have a corresponding js file
The js file must expose a run function that takes an express request and
response object. The function must return an object to be passed to the
corresponding ejs file.

No longer forcing users to be logged in.

Added a login.js file to handle the generation of the google sign in url. This
will create a different url when deployed locally for testing purposes.